+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed 
 
-- Added UDGrid Tests
+- Added UDGrid, UDLayout, UDModal, UDLink Tests
 - Fixed *-UDElement cmdlets
 - Added dependency on Universal module